refactoring code to incorporate api module
