--- public/index.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html ng-app="hey">
-<head>
-  <title>Tone Down For What</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
-  <link rel="stylesheet" type="text/css" href="css/styles.css">
-  <link href='https://fonts.googleapis.com/css?family=Lato:400,700' rel='stylesheet' type='text/css'>
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
-</head>
-<body> 
-  <button ui-sref="login" class="login-show-button" >Login</button>
-  <button ui-sref="register" >Sign Up</button>
-  <div class="container">
-  <div class="row">
-    <div class="col-lg-12">
-      <div class="content">
-        <h1>Tone Down For What</h1>
-        <h3>Web App That Uses Linguistic Analysis to Identify Tones In Text</h3>
-        <a href="#" class="action btn startBtn btn-large" >Get Started</a>
-      </div>
-    </div>
-  </div>
-</div>
-
-  
-  <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.5.3/angular.min.js"></script>
-  <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/angular-ui-router/0.2.18/angular-ui-router.min.js"></script>
-  <script type="text/javascript" src="/js/app.js"></script>
-  <script type="text/javascript" src="https://code.jquery.com/jquery-2.2.0.min.js"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
-</body>
-</html>
