+++ public/input_demo.html
@@ -1,10 +1,9 @@
 <!DOCTYPE html>
-<html ng-app="toneDown">
+<html ng-app="toneAnalyzer">
 <head>
   <title>Tone Down For What</title>
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
-<!--   <link rel="stylesheet" type="text/css" href="css/styles.css"> -->
   <link href='https://fonts.googleapis.com/css?family=Lato:400,700' rel='stylesheet' type='text/css'>
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
   <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.5.0/angular.min.js"></script>
