+++ public/partials/main_page.html
@@ -1,106 +1,22 @@
-  <div class="main_page_bkgrd">
-    <div id="wrapper"> 
-      <div id="sidebar-wrapper">
-        <ul class="sidebar-nav">
-          <li class="sidebar-brand">Welcome {{ username }}</li>
-          <li class="sidebar-brand">
-            Saved Drafts
-          </li>
-          <li>
-            <a class="draft_links" href="#">Saved Draft # 1</a>
-          </li>
-          <li>
-            <a class="draft_links" href="#">Saved Draft # 2</a>
-          </li>
-          <li>
-            <a class="draft_links" href="#">Saved Draft # 3</a>
-          </li>
-          <li>
-            <a class="draft_links" href="#">Saved Draft # 4</a>
-          </li>
-          <li>
-            <a class="draft_links" href="#">Saved Draft # 5</a>
-          </li>
-          <li>
-            <a class="draft_links" href="#">Saved Draft # 6</a>
-          </li>
-        </ul>
-      </div>
-      <div id="page-content-wrapper">
-        <div class="container-fluid">
-          <div class="row">
-            <div class="col-lg-12">
-              <h3 class="mainpge-title">Tone Down For What</h3>
-              <br/>
-            </div>
-            <div class="col-md-6 col-md-offset-3 col-xs-12">
-              <input type="text" class="form-control tone_textarea" placeholder="Title" class="validate title-text"/>
-              <br/>
-            </div>
-            <div class="col-lg-12">
-              <textarea class="form-control tone_textarea" rows="15"></textarea>  
-            </div>
-            <div class="row">
-              <div class="col-xs-12 text-center">
-                <a href="#" class="action btn analyzeBtn hover-button btn-large">Analyze</a>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
+<div ng-controller='inputForm'>
+
+  <form>
+    <textarea type='text' ng-model='toneText'></textarea>
+    <input type='submit' ng-click='analyzeTone()'>
+  </form>
+
+  <div ng-init='retrieveData()'>
+    <ul ng-repeat='(key, value) in toneDatas track by $index'>
+      <li><strong>Text to be analyzed</strong> </li>
+      <li> {{value.content}} </li>
+      <li><strong>Social tone data</strong></li>
+      <li ng-repeat='(key, value) in value.social_tone_data track by $index'> {{value.tone_type}} {{value.tone_score}} </li>
+      <li><strong>Emotional tone data</strong></li>
+      <li ng-repeat='(key, value) in value.emotion_tone_data'> {{value.tone_type}} {{value.tone_score}} </li>
+      <li><strong>Writing tone data</strong></li>
+      <li ng-repeat='(key, value) in value.writing_tone_data'> {{value.tone_type}} {{value.tone_score}} </li>
+      <draw-chart chartIndex={{$index}} ></draw-chart>
+    </ul>
   </div>
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-<!-- <div class="row">
-  <div class="col-md-3 col-xs-3 side-panel">
-    <h3>Welcome {{ username }}</h3>
-  </div>
 </div>
-    
-<div class="main_page_bkgrd">
-  <div class="container">
-    <div class="row">
-      <div class="col-md-12 col-xs-6">
-        <h3 class="mainpge-title">Tone Down For What</h3>
-        <textarea class="form-control tone_textarea" rows="15"></textarea>
-      </div>
-    </div>
-  </div>
-</div>
-
-
- -->
-
-
-<!--  <div class="container nav-container">
-   <div class="row nav-row">
-     <div class="no-float col-xs-3 side-panel">
-        <h3>Welcome {{ username }}</h3>
-     </div>
-     <div class="no-float col-xs-9 main_page_bkgrd">
-      <h3 class="mainpge-title">Tone Down For What</h3>
-      <textarea class="form-control tone_textarea" rows="15"></textarea>  
-      <div class="row">
-        <div class="col-xs-12 text-center">
-          <a href="#" class="action btn analyzeBtn hover-button btn-large">Analyze</a>
-            <a href="#menu-toggle" class="btn btn-default" id="menu-toggle">Toggle Menu</a>
-        </div>
-      </div>
-     </div>
-   </div>
- </div> -->
