+++ public/template/chartRender.html
@@ -1,0 +1,1 @@
+<div style="width:100%; height:400px;"></div>
