+++ public/views/index.html
@@ -19,5 +19,8 @@
   <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/angular-ui-router/0.2.18/angular-ui-router.min.js"></script>
   <script type="text/javascript" src="/js/app.js"></script>
   <script type="text/javascript" src="/controllers/loginController.js"></script>
+  <script type="text/javascript" src="./controllers/inputFormController.js"></script>
+  <script src="http://code.highcharts.com/highcharts.js"></script>
+  <script type="text/javascript" src="./grouped-categories.js"></script>
 </body>
 </html>
